Adds TestCase to move redundant settings handlers for reuse.
