+++ tests/Integration/inc/Engine/Optimization/Minify/JS/Subscriber/process.php
@@ -2,7 +2,7 @@
 
 namespace WP_Rocket\Tests\Integration\inc\Engine\Optimization\Minify\JS\Subscriber;
 
-use WP_Rocket\Tests\Integration\FilesystemTestCase;
+use WP_Rocket\Tests\Integration\inc\Engine\Optimization\TestCase;
 
 /**
  * @covers \WP_Rocket\Engine\Optimization\Minify\JS\Subscriber::process
@@ -17,11 +17,8 @@ use WP_Rocket\Tests\Integration\FilesystemTestCase;
  *
  * @group  MinifyJS
  */
-class Test_Process extends FilesystemTestCase {
+class Test_Process extends TestCase {
 	protected $path_to_test_data = '/inc/Engine/Optimization/Minify/JS/Subscriber/process.php';
-	protected $cnames;
-	protected $zones;
-	private   $settings;
 
 	public function setUp() {
 		$this->wp_content_dir = 'vfs://public/wordpress/wp-content';
@@ -37,7 +34,7 @@ class Test_Process extends FilesystemTestCase {
 
 		remove_filter( 'pre_get_rocket_option_minify_js', [ $this, 'return_true' ] );
 		remove_filter( 'pre_get_rocket_option_minify_js_key', [ $this, 'return_key' ] );
-		$this->unset_settings( $this->settings );
+		$this->unsetSettings();
 	}
 
 	/**
@@ -48,79 +45,11 @@ class Test_Process extends FilesystemTestCase {
 		add_filter( 'pre_get_rocket_option_minify_js_key', [ $this, 'return_key' ] );
 
 		$this->settings = $settings;
-		$this->set_settings( $settings );
+		$this->setSettings();
 
 		$this->assertSame(
 			$this->format_the_html( $expected ),
 			$this->format_the_html( apply_filters( 'rocket_buffer', $original ) )
 		);
-
-		$this->unset_settings( $settings );
-	}
-
-	private function set_settings( array $settings ) {
-		foreach ( $settings as $key => $value ) {
-			if ( 'minify_concatenate_js' === $key ) {
-				$callback = 0 === $value ? 'return_false' : 'return_true';
-				add_filter( 'pre_get_rocket_option_minify_concatenate_js', [ $this, $callback ] );
-				continue;
-			}
-
-			if ( 'cdn' === $key ) {
-				$callback = 0 === $value ? 'return_false' : 'return_true';
-				add_filter( 'pre_get_rocket_option_cdn', [ $this, $callback ] );
-				continue;
-			}
-
-			if ( 'cdn_cnames' === $key ) {
-				$this->cnames = $value;
-				add_filter( 'pre_get_rocket_option_cdn_cnames', [ $this, 'set_cnames' ] );
-				continue;
-			}
-
-			if ( 'cdn_zone' === $key ) {
-				$this->zones = $value;
-				add_filter( 'pre_get_rocket_option_cdn_zone', [ $this, 'set_zones' ] );
-				continue;
-			}
-		}
-	}
-
-	private function unset_settings( array $settings ) {
-		foreach ( $settings as $key => $value ) {
-			if ( 'minify_concatenate_js' === $key ) {
-				$callback = 0 === $value ? 'return_false' : 'return_true';
-				remove_filter( 'pre_get_rocket_option_minify_concatenate_js', [ $this, $callback ] );
-				continue;
-			}
-
-			if ( 'cdn' === $key ) {
-				$callback = 0 === $value ? 'return_false' : 'return_true';
-				remove_filter( 'pre_get_rocket_option_cdn', [ $this, $callback ] );
-				continue;
-			}
-
-			if ( 'cdn_cnames' === $key ) {
-				remove_filter( 'pre_get_rocket_option_cdn_cnames', [ $this, 'set_cnames' ] );
-				continue;
-			}
-
-			if ( 'cdn_zone' === $key ) {
-				remove_filter( 'pre_get_rocket_option_cdn_zone', [ $this, 'set_zones' ] );
-				continue;
-			}
-		}
-	}
-
-	public function return_key() {
-		return 123456;
-	}
-
-	public function set_cnames() {
-		return $this->cnames;
-	}
-
-	public function set_zones() {
-		return $this->zones;
 	}
 }
